일부 기능 분리 및 추가, 예외처리

+++ mcbot.js
@@ -1,9 +1,10 @@
 const Discord = require('discord.js');
 const fs = require("fs");
-const mysql = require('mysql');
+const con = require('./config/database')
 const bot = new Discord.Client();
 const Badwords = require("./jsons/fiterWords.json");
 const log = require('./config/logger.js');
+const swearLog = require('./config/swearLogger');
 require('dotenv').config();
 var StatsD = require('hot-shots');
 var dogstatsd = new StatsD();
@@ -11,15 +12,14 @@ var dogstatsd = new StatsD();
 bot.commands = new Discord.Collection();
 bot.aliases = new Discord.Collection();
 
+
 fs.readdir("./command/", (err, files) => {
     if (err) console.log(err);
-
     let jsfile = files.filter(f => f.split(".").pop() === "js")
     if (jsfile.length <= 0) {
         console.log("명령어를 찾지 못했어요...");
         return;
     }
-
     jsfile.forEach((f, i) => {
         let props = require(`./command/${f}`);
         console.log(`[ ${f} ] load Complete`);
@@ -30,24 +30,13 @@ fs.readdir("./command/", (err, files) => {
     });
 });
 
-const con = mysql.createConnection({
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-});
-
-con.connect(err => {
-    if (err) throw err;
-    console.log('데이터베이스 연결 완료!');
-});
 
 function generatexp() {
     let min = 1;
     let max = 5;
-
     return Math.floor(Math.random() * (max - min + 1)) + min;
 }
+
 
 bot.on('ready', () => {
     console.log(`┌────────────────────────────┐`);
@@ -61,9 +50,7 @@ bot.on('ready', () => {
         `채널 : ${bot.channels.cache.size}개`,
         `MCprefix로 서버별 접두사 확인`
     ]
-
     let value = 0;
-
     setInterval(function () {
         let status = statuses[value];
         value++;
@@ -71,11 +58,13 @@ bot.on('ready', () => {
             value = 0;
         }
         bot.user.setActivity(status, { type: "PLAYING" });
+        
     }, 3000);
-
 });
 
+
 bot.on('guildCreate', (guild) => {
+    
     log.info(`MCBOT이 새로운 길드에 초대됨. 길드 : ${guild.name}`);
     try {
         con.query(`INSERT INTO Guilds (guildId, GuildOwnerId) VALUES('${guild.id}', ${guild.ownerID});`);
@@ -85,7 +74,9 @@ bot.on('guildCreate', (guild) => {
     }
 });
 
+
 bot.on('guildDelete', async (guild) => {
+    
     log.info(`MCBOT이 길드에서 퇴장됨. 길드 : ${guild.name}`);
     try {
         await con.query(`DELETE FROM Guilds WHERE guildId = '${guild.id}';`);
@@ -95,6 +86,7 @@ bot.on('guildDelete', async (guild) => {
         log.error(`Error while exiting guild ${err}`)
     }
 })
+
 
 bot.on('message', async message => {
 
@@ -139,21 +131,45 @@ bot.on('message', async message => {
             memberCountChannel.setName(`${ChName} : ` + memberCount);
         }
     });
-
+    //─────────────────────────────────────────────────────────────────────────────────────────────────────────────
     // 욕설 필터링
-    for (var i in filterwords) {
-        if (message.content.toLowerCase().includes(filterwords[i].toLowerCase())) {
-            foundText = true;
-        }
-    }
-
-    for (var k in messageURL) {
-        if (message.content.toLowerCase().includes(messageURL[k].toLowerCase())) {
-            foundText = false;
-        }
-    }
-
     let messagechid = message.channel.id;
+
+    function filterLog(find) {
+        if (find) {
+            con.query(`SELECT * FROM Guilds WHERE guildId = '${message.guild.id}';`, (err, rows) => {
+                if (err) throw err;
+                let exceptionCh = rows[0].exceptionCh;
+                let logCh = rows[0].logCh;
+                if (!exceptionCh) {
+                    return;
+                }
+                if (messagechid === exceptionCh) {
+                    return;
+                } else {
+                    swearLog.info(`\`${message.guild.name}\`의 '${message.channel.name}'에서 '${message.author.username}'님이 단어 '${message.content}'를(을) 사용하여 필터링 됨`)
+
+                    let ch = bot.channels.cache.get(`${logCh}`);
+
+                    if (!logCh) {
+                        message.delete();
+                        return message.reply("(이쁜말)");
+                    } else {
+                        let embed = new Discord.MessageEmbed()
+                            .setTitle("욕설 필터링")
+                            .setDescription(`\`${message.author.username}\`님이 욕설을 사용하였습니다.`)
+                            .setColor("RED")
+                            .setTimestamp()
+                            .addField("[ 사용 단어 ]", `${message.content}`)
+                            .addField("[ 태그 ]", `<@${message.author.id}>`)
+                        ch.send(embed);
+                        message.delete();
+                        return message.reply("(이쁜말)");
+                    }
+                }
+            });
+        }
+    }
 
     if (message.guild.id === "534586842079821824") {
         for (var i in zbcFilter) {
@@ -162,47 +178,41 @@ bot.on('message', async message => {
             }
         }
 
-        if (foundText) {
-            con.query(`SELECT * FROM Guilds WHERE guildId = '${message.guild.id}';`, (err, rows) => {
-                if (err) throw err;
-                let exceptionCh = rows[0].exceptionCh;
-                if (!exceptionCh) {
-                    return;
-                }
-                if (messagechid === exceptionCh) {
-                    return;
-                } else {
-                    message.delete();
-                    return message.reply("(이쁜말)");
-                }
-            });
+        for (var k in messageURL) {
+            if (message.content.toLowerCase().includes(messageURL[k].toLowerCase())) {
+                foundText = false;
+            }
         }
+
+        filterLog(foundText);
+
     } else {
-        if (foundText) {
-            con.query(`SELECT * FROM Guilds WHERE guildId = '${message.guild.id}';`, (err, rows) => {
-                if (err) throw err;
-                let exceptionCh = rows[0].exceptionCh;
-                if (!exceptionCh) {
-                    return;
-                }
-                if (messagechid === exceptionCh) {
-                    return;
-                } else {
-                    message.delete();
-                    return message.reply("(이쁜말)");
-                }
-            });
+        for (var i in filterwords) {
+            if (message.content.toLowerCase().includes(filterwords[i].toLowerCase())) {
+                foundText = true;
+            }
         }
+
+        for (var k in messageURL) {
+            if (message.content.toLowerCase().includes(messageURL[k].toLowerCase())) {
+                foundText = false;
+            }
+        }
+
+        filterLog(foundText);
 
     }
-
+    //─────────────────────────────────────────────────────────────────────────────────────────────────────────────
 
     //xp 시스템
     con.query(`SELECT * FROM xp WHERE guildId = '${message.guild.id}' AND id = '${message.author.id}';`, (err, rows) => {
         if (err) throw err;
 
         if (rows.length < 1) {
-            con.query(`INSERT INTO xp (guildId, id, xp, name) VALUES ('${message.guild.id}', '${message.author.id}', ${generatexp()}, '${message.author.username}');`);
+            let name = message.author.username;
+            name = name.replace(`'`, "\\'");
+            name = name.replace(`"`, '\\"');
+            con.query(`INSERT INTO xp (guildId, id, xp, name) VALUES ('${message.guild.id}', '${message.author.id}', ${generatexp()}, '${name}');`);
         } else {
             let xp = rows[0].xp;
             let lvl = rows[0].lvl;
@@ -250,7 +260,12 @@ bot.on('message', async message => {
     }
 });
 
+
 bot.on('guildMemberAdd', member => {
+
+    if (member.guild.id === "534586842079821824") {
+        return;
+    }
     let Guild = member.guild.id;
     con.query(`SELECT * FROM Guilds WHERE guildId = '${member.guild.id}';`, (err, rows) => {
         let logCh = rows[0].logCh;
@@ -298,7 +313,12 @@ bot.on('guildMemberAdd', member => {
     });
 });
 
+
 bot.on('guildMemberRemove', member => {
+
+    if (member.guild.id === "534586842079821824") {
+        return;
+    }
     let Guild = member.guild.id;
     con.query(`DELETE FROM xp WHERE id = '${member.id}';`);
     con.query(`SELECT * FROM Guilds WHERE guildId = '${member.guild.id}';`, (err, rows) => {
@@ -335,19 +355,40 @@ bot.on('guildMemberRemove', member => {
         } else {
             memberCountChannel.setName(`${ChName} : ` + memberCount);
         }
-
     });
 });
 
+
 bot.on('guildMemberUpdate', member => {
-    let name = member.user.username
-    if(name.includes("'")) {
-        name.replace("'", "\'")
-    } else if (name.includes('"')) {
-        name.replace('"', '\"')
+
+    let memid = member.id;
+    let memname = member.user.username;
+
+    memname = memname.replace(`'`, "\\'");
+    memname = memname.replace(`"`, '\\"');
+    try {
+        con.query(`SELECT * FROM xp WHERE id = '${memid}';`, (err, rows) => {
+            if (err) throw err;
+            if (rows.length < 1) {
+                return;
+            } else {
+                con.query(`UPDATE xp Set name = \'${memname}\' WHERE id = '${memid}';`);
+                return;
+            }
+        })
+        con.query(`SELECT * FROM warnUser WHERE Id = '${memid}';`, (err, rows) => {
+            if (err) throw err;
+            if (rows.length < 1) {
+                return;
+            } else {
+                con.query(`UPDATE warnUser SET name = \'${memname}\' WHERE Id = '${memid}';`);
+                return;
+            }
+        })
+    } catch (error) {
+        return error;
     }
-    con.query(`UPDATE xp Set name = "${name}" WHERE id = "${member.id}";`);
-    con.query(`UPDATE warnUser SET name = "${name}" WHERE id = "${member.id}";`);
 });
+
 
 bot.login(process.env.MCBOT_TOKEN);
